requirements.txt
python-dotenv~=1.0.0
pinecone-client~=2.2.2
langchain~=0.0.194
